Extracted tests requiring standalone Ray cluster into a standalone module (to make sure these are not affecting module-scoped, shared fixtures, and therefore other tests in the same module)
